runtime,syscall,link: add GOOS=hurd GOARCH=amd64 support
Port the GNU/Hurd Go runtime, syscall tables and linker/obj support to
amd64, alongside the existing hurd/386 port:

  - runtime: defs_hurd_amd64.go, signal_hurd_amd64.go, rt0/sys asm with
    the System V AMD64 ABI libc trampolines; share the OS constants and
    split the i386-specific types.  glibc owns the TLS base (%fs), so
    settls is a no-op and the asm_amd64.s TLS setup is skipped.  The
    signal trampoline uses the ABI0 handler like 386/Solaris.
  - syscall: amd64 generated tables (zerrors/ztypes/zsyscall/dump),
    asm_hurd_amd64.s data slots and the ABI0 syscall6/rawSyscall6 stubs.
  - linker/obj: register Hhurd on amd64 (TLS base, CanUse1InsnTLS,
    self-contained ELF interpreter, archinit) and emit R_X86_64_64
    dynamic relocations for libc import data slots.

Commit as David Bidner <david@crap.solutions>.

+++ src/syscall/syscall_hurd_amd64.go
@@ -2,12 +2,12 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-//go:build hurd && 386
+//go:build hurd && amd64
 
 package syscall
 
 func (iov *Iovec) SetLen(length int) {
-	iov.Len = uint32(length)
+	iov.Len = uint64(length)
 }
 
 func (msghdr *Msghdr) SetControllen(length int) {
@@ -19,9 +19,9 @@ func (cmsg *Cmsghdr) SetLen(length int) {
 }
 
 func setTimespec(sec, nsec int64) Timespec {
-	return Timespec{Sec: int32(sec), Nsec: int32(nsec)}
+	return Timespec{Sec: sec, Nsec: nsec}
 }
 
 func setTimeval(sec, usec int64) Timeval {
-	return Timeval{Sec: int32(sec), Usec: int32(usec)}
+	return Timeval{Sec: sec, Usec: usec}
 }
